Fix: DPL: proactively set known limit on inverters
when starting up, the current limit set at the inverter is not known for
a couple of minutes (see upstream FAQ). to speed things up and to make
sure that the current limit is eventually known (which it is not always,
see #1427), the DPL now sends the lower power limit proactively.