BREAKING CHANGE: huawei grid charger with MCP2515 requires IRQ pin
it is likely that the previous implementation for the MCP2515 CAN
controller and transceiver communicating with the Huawei charger did
work even without the IRQ pin in use. the new implementation will miss
CAN messages if the IRQ pin is not connected or not defined, because it
is event-driven, rather than polling with maximum frequency.

to avoid people yelling at us because their insane setup worked before
but not does not any more, we mark this to be a breaking change.