Fix: allow management of DPL inverters while DPL is off
we do allow to send limit updates and to start or stop the inverter
using the web UI or MQTT payloads or any other supported mean even if
this inverter is governed by the DPL, while the DPL is disabled. this
is also documented explicitly (see DPL mode MQTT docs).

after implementing support for multiple inverters, this behavior changed
unintentionally. this changeset makes sure that inverters are shut down
once when the DPL dis disabled, but are left alone afterwards for as
long as the DPL is disabled.