Fix: really trigger standby only once on DPL shutdown
relying on updateInverters() to return false after requesting inverter
standby is not a good idea. the standby function will always set a new
limit, and the updateInverters() function may fail to recognize that no
change is actually required.

instead, we trigger the standby only once and rely on the
updateInverters() function being called as part of the DPL loop() to
apply the standby state.