Fix: flush Serial after each line
the serial output was still mangled from time to time for no apparent
reason. in particular, when dumping R48xx board properties, which is a
"large" multi-line string, almost always characters were missing.

when using Serial.flush() after writing each line, the issue is
resolved. so we now spend the time to make sure the Serial buffer is
flushed after each line. this should keep the issue resolved until lines
itself are too long, in which case the output shall be re-formatted as
the line will be unreasonably long.