Fix: add smart buffer output to total DPL output
to keep the DPL icon in the web UI in the header of VE.Direct MPPT
charge controllers live view alive, which tells us how much power the
DPL is requesting from the governed inverters, we set a variable which
adds up the power covered by the respective types of inverters. when
support for smart-buffer-powered inverters arrived, we forgot to add
their output to this value.