move info about inverter output metering to inverter dump
we will know what each governed inverter is producing if verbose logging
is enabled already at the end of each DPL loop, where all info
concerning all DPL inverters is dumped. however, we shall make the info
available there whether or not the inverter is configured as being
"behing" the power meter, i.e., whether or not its output is part of the
power meter reading.